+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-progress
 4.1.12 (unreleased)
 -------------------
 
+- Catch and pass missing coordinates error when creating location from measurement.
+
 - Avoid running calc_progress in projectreview view, show reviewproject.progress instead.
 
 - Disable Download reviews button if there are no reviews.