+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog of lizard-progress
 ===================================================
 
-4.1.12 (unreleased)
+4.2.0 (2018-07-18)
 
 - Added work_impossible test before reporting error on missing start point